REFACTOR: upload preprocessor classes and Gradle build config
