+++ imagekit/build.gradle
@@ -4,6 +4,7 @@ apply plugin: 'kotlin-android'
 
 apply plugin: 'kotlin-kapt'
 apply plugin: 'maven-publish'
+apply plugin: 'jacoco'
 
 group = 'com.github.imagekit-developer'
 
@@ -67,7 +68,7 @@ dependencies {
 
     api 'io.reactivex.rxjava2:rxandroid:2.1.1'
 
-    api "com.google.dagger:dagger:$dagger_version"
+    implementation "com.google.dagger:dagger:$dagger_version"
     kapt "com.google.dagger:dagger-compiler:$dagger_version"
     api "org.jetbrains.kotlin:kotlin-reflect:1.3.50"
 
@@ -76,36 +77,56 @@ dependencies {
     implementation 'com.squareup.okhttp3:logging-interceptor:3.12.0'
     implementation 'com.squareup.retrofit2:adapter-rxjava2:2.4.0'
     testImplementation 'com.squareup.okhttp3:mockwebserver:3.12.0'
-    compileOnly 'com.github.pengrad:jdk9-deps:1.0'
+//    compileOnly 'com.github.pengrad:jdk9-deps:1.0'
 
     implementation "com.linkedin.android.litr:litr:1.4.16"
 }
 
 apply plugin: 'org.jetbrains.dokka'
-apply plugin: 'jacoco-android'
+//
+
+jacoco {
+    toolVersion "0.8.8"
+}
 
 tasks.withType(Test) {
     jacoco.includeNoLocationClasses = true
+    finalizedBy jacocoTestReport
 }
 
-jacocoAndroidUnitTestReport {
-    csv.enabled false
-    html.enabled true
-    xml.enabled true
-    excludes += ['**/R.class',
-                 '**/BR.class',
-                 '**/R$*.class',
-                 '**/*$ViewInjector*.*',
-                 '**/*$ViewBinder*.*',
-                 '**/BuildConfig.*',
-                 'android/**',
-                 '**/Manifest*.*',
-                 '**/*$Lambda$*.*', // Jacoco can not handle several "$" in class name.
-                 '**/*Module.*', // Modules for Dagger.
-                 '**/*Dagger*.*', // Dagger auto-generated code.
-                 '**/*MembersInjector*.*', // Dagger auto-generated code.
-                 '**/*_Provide*Factory*.*',
-                 '**/*_Factory.*', //Dagger auto-generated code
-                 '**/*$*$*.*' // Anonymous classes generated by kotlin
-    ]
+tasks.register('jacocoTestCoverageVerification', JacocoCoverageVerification) {
+    violationRules {
+        rule {
+            enabled = true
+            element = 'CLASS'
+
+        }
+    }
+}
+
+task jacocoTestReport(type: JacocoReport, dependsOn: ['testDebugUnitTest']) {
+    reports {
+        csv.enabled false
+        html.enabled true
+        xml.enabled true
+    }
+    classDirectories.setFrom(files(classDirectories.files.collect {
+        fileTree(dir: it, excludes: [
+                '**/R.class',
+                '**/BR.class',
+                '**/R$*.class',
+                '**/*$ViewInjector*.*',
+                '**/*$ViewBinder*.*',
+                '**/BuildConfig.*',
+                'android/**',
+                '**/Manifest*.*',
+                '**/*$Lambda$*.*', // Jacoco can not handle several "$" in class name.
+                '**/*Module.*', // Modules for Dagger.
+                '**/*Dagger*.*', // Dagger auto-generated code.
+                '**/*MembersInjector*.*', // Dagger auto-generated code.
+                '**/*_Provide*Factory*.*',
+                '**/*_Factory.*', //Dagger auto-generated code
+                '**/*$*$*.*' // Anonymous classes generated by kotlin
+        ])
+    }))
 }
