+++ sample/src/main/java/com/imagekit/android/sample/MainActivity.kt
@@ -5,22 +5,25 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.imagekit.android.ImageKit
 import com.imagekit.android.entity.TransformationPosition
-import kotlinx.android.synthetic.main.activity_main.*
+import com.imagekit.android.sample.databinding.ActivityMainBinding
+
+//import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        val binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         ImageKit.init(
             context = applicationContext,
-            publicKey = "YOUR_PUBLIC_KEY",
-            urlEndpoint = "https://ik.imagekit.io/YOUR_IMAGEKIT_ID",
+            publicKey = "public_5P5QM23aRv9XkOcfJO1okZ0DzOw=",
+            urlEndpoint = "https://ik.imagekit.io/tqhfz73me",
             transformationPosition = TransformationPosition.PATH,
-            authenticationEndpoint = "YOUR_AUTHENTICATION_ENDPOINT"
+            authenticationEndpoint = "https://2303-2401-4900-1c46-e905-38ef-8dd2-56ae-1080.ngrok-free.app/"
         )
 
-        btUploadImage.setOnClickListener {
+        binding.btUploadImage.setOnClickListener {
             startActivity(
                 Intent(
                     this@MainActivity,
@@ -29,7 +32,7 @@ class MainActivity : AppCompatActivity() {
             )
         }
 
-        btUploadFile.setOnClickListener {
+        binding.btUploadFile.setOnClickListener {
             startActivity(
                 Intent(
                     this@MainActivity,
@@ -38,7 +41,7 @@ class MainActivity : AppCompatActivity() {
             )
         }
 
-        btUrlConstruct.setOnClickListener {
+        binding.btUrlConstruct.setOnClickListener {
             startActivity(
                 Intent(
                     this@MainActivity,
